ENH: Rename MONAILabelReviewer helper library
Slicer standard is to name the corresponding lib {ModuleName}Lib so there is no potential conflict of scripted module libraries which are located in the same scripted module directory.
